Closes #2
The issue was that the used getExternal(rfid) method had the additional sql condition
"OCPP_TAG.BLOCKED.isFalse()", which is fine if you want to find out external and _active_
(read: verified, not blocked) tags for normal operations. however, this condition causes
problems when we want to verify the rfid @ ps. it should have been omitted.